/store -> shop refactoring
