+++ hamza-client/src/modules/shop/components/product-card.tsx
