+++ hamza-client/src/modules/shop/components/range-slider.tsx
