+++ hamza-client/src/modules/shop/data/category-icons.ts
