+++ hamza-client/src/modules/shop/templates/paginated-products.tsx
@@ -1,7 +1,7 @@
 import { getProductsListWithSort, getRegion } from '@lib/data';
 import ProductPreview from '@modules/products/components/product-preview';
-import { Pagination } from '@modules/store/components/pagination';
-import { SortOptions } from '@modules/store/components/refinement-list/sort-products';
+import { Pagination } from '@modules/shop/components/pagination';
+import { SortOptions } from '@modules/shop/components/refinement-list/sort-products';
 
 const PRODUCT_LIMIT = 12;
